Fixed gravity, added collisions with objects
* Fixed gravity
* Converted all object values to [ float ]
+ Added multi object support
+ Added collisions with objects
+ Added lose of speed when collided

Gravity is set to 0.5f per draw
When you collide, you lose 10% of you speed
You can now collide with other objects! (kind of)
Positioning of new objects is now random

+++ Game engine/Game engine/FormObject.cs
@@ -11,20 +11,20 @@ namespace Game_engine
     class FormObject
     {
         // dx, dy jsou zmeny x, y, ktery se meni pri kazdem vykresleni
-        public int x, y, dx, dy, length, width;
+        public float x, y, dx, dy, length, width;
         public bool active, gravity, ghost;
         public Image texture;
 
 
         /// <param name="ghost">If ghost, it is not affected by any collision</param>
-        public FormObject(int x, int y, int width, int length, bool gravity, bool ghost)
+        public FormObject(float x, float y, float width, float length, bool gravity, bool ghost)
         {            
             this.x = x;
             this.y = y;
             this.width = width;
             this.length = length;
-            dx = 0;
-            dy = 0;
+            dx = 0f;
+            dy = 0f;
             active = true;
             this.gravity = gravity;
             this.ghost = ghost;
@@ -34,14 +34,14 @@ namespace Game_engine
 
         
         /// <param name="ghost">If ghost, it is not affected by any collision</param>
-        public FormObject(int x, int y, int width, int length, bool gravity, bool ghost, Image texture)
+        public FormObject(float x, float y, float width, float length, bool gravity, bool ghost, Image texture)
         {
             this.x = x;
             this.y = y;
             this.width = width;
             this.length = length;
-            dx = 0;
-            dy = 0;
+            dx = 0f;
+            dy = 0f;
             active = true;
             this.gravity = gravity;
             this.ghost = ghost;
